refatoring and add firebase
