+++ lib/app/home/transactions_page.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/material.dart';
+import 'package:wallet_manager/app/home/widgets/billing_item_widget.dart';
 import 'package:wallet_manager/models/bank_account.dart';
 import 'package:wallet_manager/shared/main_stances.dart';
-import 'package:wallet_manager/view/home/widgets/billing_item_widget.dart';
+
 
 import '../../models/transaction.dart';
