+++ lib/app/home/widgets/middle_info.dart
