+++ lib/app/home/widgets/percentage_related.dart
@@ -1,7 +1,9 @@
 
 
 import 'package:flutter/material.dart';
-import 'package:wallet_manager/view/styles/container_decorators.dart';
+
+
+import '../../styles/container_decorators.dart';
 
 class PercentageRelated extends StatelessWidget {
   const PercentageRelated({super.key});
