+++ lib/app/login/login_build.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+
+
+import 'login_page.dart';
+
+class LoginBuild extends StatelessWidget {
+  const LoginBuild({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const LoginPage();
+  }
+}
