+++ lib/controller/login_controller.dart
@@ -17,26 +17,7 @@ class LoginController {
   });
 
   login() async {
-    message.value = '';
-    String? result = await helper.getData('users');
-    if (result != null) {
-      List<User> users = _fromStringList(result);
-      for (int i = 0; i < users.length; i++) {
-        if (users[i].email == user.email &&
-            users[i].password == user.password) {
-          return true;
-        }
-      }
-    }
-    message.value = 'Email or password is not correct';
+
   }
 
-  List<User> _fromStringList(String list) {
-    List<dynamic> resultList = jsonDecode(list);
-    List<User> users = [];
-    for (Map<String,dynamic> json in resultList) {
-      users.add(User.fromMap(json));
-    }
-    return users;
-  }
 }
