+++ lib/services/auth/auth_service_helper.dart
@@ -1,0 +1,5 @@
+import '../../models/user.dart';
+
+abstract class AuthServiceHelper {
+  Future<User>  auth();
+}
