+++ pubspec.yaml
@@ -41,6 +41,7 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
   flutter_pluggy_connect: ^1.3.0
+  google_sign_in: ^5.0.0
 
 dev_dependencies:
   flutter_test:
